add if -f and ! -L file

+++ makesymlinks.sh
@@ -13,7 +13,7 @@ files="vimrc"    # list of files/folders to symlink in homedir
 ##########
 
 # create dotfiles_old in homedir
-echo "Creating $olddir for backup of any existing dotfiles in $(~)"
+echo "Creating $olddir for backup of any existing dotfiles in $(HOME)"
 mkdir -p $olddir
 echo "...done"
 
@@ -24,8 +24,10 @@ echo "...done"
 
 # move any existing dotfiles in homedir to dotfiles_old directory, then create symlinks 
 for file in $files; do
-    echo "Moving any existing dotfiles from $(~) to $olddir"
-    mv ~/.$file ~/dotfiles_old/
+    if [[ -f "${HOME}/.${file}" && ! -L "${HOME}/.${file}" ]];then
+        echo "Moving existing dotfile \".${file}\" from ${HOME} to $olddir"
+        mv ~/.$file ~/dotfiles_old/
+    fi
     echo "Creating symlink to $file in home directory."
     ln -s $dir/$file ~/.$file
 done
